requirements.txt
Django[argon2]==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
dj_database_url==2.3.0
PyJWT==2.9.0
django-cors-headers==4.6.0
gunicorn==23.0.0
uvicorn
django-redis